pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic==1.14.0",
    "asyncpg>=0.30.0",
    "email-validator>=2.0.0",
    "fastapi==0.115.0",
    "passlib[bcrypt]==1.7.4",
    "pgvector==0.2.5",
    "pydantic-settings==2.2.1",
    "pypdf2==3.0.1",
    "python-dotenv==1.0.1",
    "python-jose[cryptography]==3.3.0",
    "python-multipart==0.0.9",
    "sqlalchemy>=2.0.40",
    "uvicorn[standard]==0.32.0",
]
